+++ src/main/scala/io/viash/schemas/ParameterSchema.scala
@@ -15,42 +15,15 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package io.viash.helpers
+package io.viash.schemas
 
 import scala.reflect.runtime.universe._
-import io.circe.{Printer => JsonPrinter}
-import io.circe.syntax.EncoderOps
-import io.viash.helpers.Circe._
-
-import io.circe.Encoder
-import io.circe.generic.extras.semiauto.deriveConfiguredEncoder
-import io.viash.functionality.Functionality
-import io.viash.platforms._
-import io.viash.platforms.requirements._
-
-final case class CollectedSchemas (
-  // functionality
-  functionality: List[ParameterSchema],
-  // platforms
-  nativePlatform: List[ParameterSchema],
-  dockerPlatform: List[ParameterSchema],
-  nextflowLegacyPlatform: List[ParameterSchema],
-  // platform requirements
-  apkRequirements: List[ParameterSchema],
-  aptRequirements: List[ParameterSchema],
-  dockerRequirements: List[ParameterSchema],
-  javascriptRequirements: List[ParameterSchema],
-  pythonRequirements: List[ParameterSchema],
-  rRequirements: List[ParameterSchema],
-  rubyRequirements: List[ParameterSchema],
-  yumRequirements: List[ParameterSchema],
-)
 
 final case class ParameterSchema(
   name: String,
   `type`: String,
   description: Option[String],
-  example: List[ExampleSchema],
+  example: Option[List[ExampleSchema]],
   since: Option[String],
   deprecated: Option[DeprecatedOrRemovedSchema],
   removed: Option[DeprecatedOrRemovedSchema],
@@ -106,10 +79,15 @@ object ParameterSchema {
 
     val description = annStrings.collectFirst({case (name, value) if name.endsWith("description") => value.head})
     val example = annStrings.collect({case (name, value) if name.endsWith("example") => value}).map(ExampleSchema(_))
+    val exampleWithDescription = annStrings.collect({case (name, value) if name.endsWith("exampleWithDescription") => value}).map(ExampleSchema(_))
+    val examples = example ::: exampleWithDescription match {
+      case l if l.length > 0 => Some(l)
+      case _ => None
+    }
     val since = annStrings.collectFirst({case (name, value) if name.endsWith("since") => value.head})
     val deprecated = annStrings.collectFirst({case (name, value) if name.endsWith("deprecated") => value}).map(DeprecatedOrRemovedSchema(_))
     val removed = annStrings.collectFirst({case (name, value) if name.endsWith("removed") => value}).map(DeprecatedOrRemovedSchema(_))
-    ParameterSchema(name_, `type`, description, example, since, deprecated, removed)
+    ParameterSchema(name_, `type`, description, examples, since, deprecated, removed)
   }
 }
 
@@ -127,59 +105,15 @@ object DeprecatedOrRemovedSchema {
 final case class ExampleSchema(
   example: String,
   format: String,
+  description: Option[String],
 )
 
 object ExampleSchema {
   def apply(l: List[String]): ExampleSchema = {
-    ExampleSchema(l(0), l(1))
-  }
-}
-
-object CollectedSchemas {
-  private val jsonPrinter = JsonPrinter.spaces2.copy(dropNullValues = true)
-
-  private implicit val encodeConfigSchema: Encoder.AsObject[CollectedSchemas] = deriveConfiguredEncoder
-  private implicit val encodeParameterSchema: Encoder.AsObject[ParameterSchema] = deriveConfiguredEncoder
-  private implicit val encodeDeprecatedOrRemoved: Encoder.AsObject[DeprecatedOrRemovedSchema] = deriveConfiguredEncoder
-  private implicit val encodeExample: Encoder.AsObject[ExampleSchema] = deriveConfiguredEncoder
-
-  private def trimTypeName(s: String) = {
-    // first: io.viash.helpers.Circe.OneOrMore[String] -> OneOrMore[String]
-    // second: List[io.viash.platforms.requirements.Requirements] -> List[Requirements]
-    s.replaceAll("^(\\w*\\.)*", "").replaceAll("""(\w*)\[[\w\.]*?([\w,]*)(\[_\])?\]""", "$1 of $2")
-  }
-
-  private def annotationsOf[T: TypeTag]() = {
-    val annMembers = typeOf[T].members.map(x => (x.fullName, x.info.toString(), x.annotations)).filter(_._3.length > 0)
-    val annThis = ("__this__", typeOf[T].typeSymbol.name.toString(), typeOf[T].typeSymbol.annotations)
-    val allAnnotations = annThis :: annMembers.toList
-    // filter out any information not from our own class and lazy evaluators (we'll use the standard one - otherwise double info and more complex)
-    allAnnotations
-      .filter(a => a._1.startsWith("io.viash") || a._1 == "__this__")
-      .filter(!_._2.startsWith("=> "))
-      .map({case (a, b, c) => (a, trimTypeName(b), c)})
-  }
-
-  private def getSchema[T: TypeTag] = {
-      annotationsOf[T].map({case (a, b, c) => ParameterSchema(a, b, c)})
-  }
-
-  def export() {
-    val data = CollectedSchemas(
-      functionality          = getSchema[Functionality],
-      nativePlatform         = getSchema[NativePlatform],
-      dockerPlatform         = getSchema[DockerPlatform],
-      nextflowLegacyPlatform = getSchema[NextflowLegacyPlatform],
-      apkRequirements        = getSchema[ApkRequirements],
-      aptRequirements        = getSchema[AptRequirements],
-      dockerRequirements     = getSchema[DockerRequirements],
-      javascriptRequirements = getSchema[JavaScriptRequirements],
-      pythonRequirements     = getSchema[PythonRequirements],
-      rRequirements          = getSchema[RRequirements],
-      rubyRequirements       = getSchema[RubyRequirements],
-      yumRequirements        = getSchema[YumRequirements],
-    )
-    val str = jsonPrinter.print(data.asJson)
-    println(str)
+    l match {
+      case _ if l.length == 2 => ExampleSchema(l(0), l(1), None)
+      case _ if l.length == 3 => ExampleSchema(l(0), l(1), Some(l(2)))
+    }
+    
   }
 }