+++ src/test/scala/io/viash/MainBuildNativeSuite.scala
@@ -6,7 +6,7 @@ import java.nio.file.Paths
 import io.viash.config.Config
 
 import scala.io.Source
-import io.viash.helpers._
+import io.viash.helpers.{IO, Exec}
 
 class MainBuildNativeSuite extends FunSuite with BeforeAndAfterAll {
   // which platform to test