+++ src/test/scala/io/viash/auxiliary/MainBuildAuxiliaryNativeParameterCheck.scala
@@ -6,7 +6,7 @@ import java.nio.file.{Paths, Files, StandardCopyOption}
 import io.viash.config.Config
 
 import scala.io.Source
-import io.viash.helpers._
+import io.viash.helpers.{IO, Exec}
 import io.viash.TestHelper
 
 class MainBuildAuxiliaryNativeParameterCheck extends FunSuite with BeforeAndAfterAll {