+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## NEW FUNCTIONALITY
 
-* Allow setting the number of processes and memory limit from within the Viash config. Example:
+* Allow setting the number of processes and memory limit from within the Viash config, 
+  as well as a list of required commands. Example:
 
   ```yaml
   functionality:
@@ -10,6 +11,7 @@
   requirements:
     n_proc: 10
     memory: 10G
+    commands: [ bash, r, perl ]
   ```
   
   You can override the default requirements at runtime:
@@ -68,6 +70,10 @@
 
 * `MainTestDockerSuite`: Derive config alternatives from the base `vsh.yaml` instead of adding the changes in separate files.
   This both reduces file clutter and prevents having to change several files when there are updates in the config format.
+
+# BUG FIXES
+
+* `NextflowVdsl3Platform`: Don't generate an error when `--publish_dir` is not defined and `-profile no_publish` is used.
 
 # Viash 0.5.15